qPCR_run1 {EasyqpcR}	R Documentation

Raw data from the first qPCR run.

Description

This is a dataset containg the raw data (Cq values) of the first qPCR run (the first biological replicate).

Usage

data(qPCR_run1)

Format

A data frame with 15 observations on the following 5 variables.

Samples

a factor with levels Calibrator Control 1 Control 2 Treatment 1 Treatment 2

RG1

a numeric vector

RG2

a numeric vector

TG

a numeric vector

TGb

a numeric vector

Details

This data.frame is composed by 2 reference genes (RG1 and RG2), and two target genes (TG and TGb). There are 5 samples, starting by the controls (1 and 2), the treatment groups (1 and 2), and a calibrator (calibrator). It is the first biological replicate and there is a qPCR triplicate by sample.

Source

S. Le Pape, IRTOMIT-INSERM U1082.

References

S. Le Pape, IRTOMIT-INSERM U1082.

Examples

data(qPCR_run1)

[Package EasyqpcR version 1.28.0 Index]