slope {EasyqpcR}	R Documentation

Function to calculate the amplification efficiency

Description

This function calculates the amplification efficiency from classical qPCR dilution experiment using the Cq values. The Cq values are plotted against the logarithmized concentration (or dilution) values, a linear regression line is fit and the efficiency calculated by E = 10^-1/slope.

Usage

slope(data, q, r, na.rm = FALSE)

Arguments

data
data.frame containing row datas (genes in columns, samples in rows, Cq values).

q
numeric, cDNA dilution values.

r
numeric, number of qPCR replicates.

na.rm
logical, indicating whether NA values should be stripped before the computation proceeds.

Value

Efficiency
Primer amplification efficiency.

Slope
Slope of the dilution curve.

Intercept
Intercept of the dilution curve

Author(s)

Sylvain Le pape <sylvain.le.pape@univ-poitiers.fr>

See Also

You can also see the qpcR package with the calib and calib2 functions.

Examples

data(Efficiency_calculation)

slope(data = Efficiency_calculation, q=c(1000, 100 ,10, 1, 0.1), r=3, na.rm=TRUE)

[Package EasyqpcR version 1.28.0 Index]